+++ assets/karma.conf.js
@@ -1,14 +1,7 @@
-// Karma configuration
-// Generated on Wed Dec 27 2017 10:17:29 GMT-0600 (CST)
-var proxyquire = require('proxyquireify');
-var browserifyBabalIstanbul = require('browserify-babel-istanbul');
-var isparta = require('isparta');
 
-function isDebug(argument) {
-    return argument === '--debug';
-}
-
-
+/**
+ * Karma configuration for WDFN assets
+ */
 
 module.exports = function (config) {
     /**
@@ -20,12 +13,13 @@ module.exports = function (config) {
 
         // frameworks to use
         // available frameworks: https://npmjs.org/browse/keyword/karma-adapter
-        frameworks: ['browserify', 'jasmine-ajax', 'jasmine'],
+        frameworks: ['jasmine-ajax', 'jasmine'],
 
         // list of files / patterns to load in the browser
         files: [
             'tests/scripts/globalConfig.js',
             'src/scripts/**/*.js'
+            //'src/scripts/store/index.spec.js'
         ],
 
         // list of files / patterns to exclude
@@ -36,16 +30,17 @@ module.exports = function (config) {
         // preprocess matching files before serving them to the browser
         // available preprocessors: https://npmjs.org/browse/keyword/karma-preprocessor
         preprocessors: {
-            'src/scripts/**/*.js': ['browserify']
+            'src/scripts/**/*.js': ['webpack', 'sourcemap']
         },
 
-        browserify: {
-            debug: process.argv.some(isDebug),
-            configure: function (bundle) {
-                bundle
-                    .plugin(proxyquire.plugin)
-                    .require(require.resolve('./src/scripts'), {entry: true});
-            }
+        webpack: {
+            ...require('./webpack.config'),
+            mode: 'development',
+            devtool: 'inline-source-map'
+        },
+        webpackMiddleware: {
+            // webpack-dev-middleware configuration
+            stats: 'errors-only'
         },
 
         // test results reporter to use
@@ -86,8 +81,8 @@ module.exports = function (config) {
         karmaConfig = {
             ...karmaConfig,
             reporters: [
-                ...karmaConfig.reporters,
-                'coverage'
+                ...karmaConfig.reporters
+                //'coverage'
             ],
 
             coverageReporter: {
@@ -96,15 +91,6 @@ module.exports = function (config) {
                     {type: 'cobertura', dir: 'coverage/'},
                     {type: 'lcovonly', dir: 'coverage/'}
                 ]
-            },
-
-            browserify: {
-                ...karmaConfig.browserify,
-                transform: [browserifyBabalIstanbul({
-                    instrumenter: isparta,
-                    instrumenterConfig: {babel: {presets: ['env']}},
-                    ignore: ['**/lib/**', '**/*.spec.js']
-                })]
             }
         };
     } else {
@@ -137,7 +123,7 @@ module.exports = function (config) {
                 // IE 11 failing with timezone issues
                 //'bs_ie11_windows10',
                 'bs_chrome52_windows10',
-                'bs_firefox52_windows10',
+                'bs_firefox52_windows10'
                 // Galaxy browser times out trying to connect
                 //'bs_galaxys8_chrome52'
             ]