+++ assets/src/scripts/components/hydrograph/layout.spec.js
@@ -1,22 +1,26 @@
-const proxyquire = require('proxyquireify')(require);
-const { format } = require('d3-format');
+import { format } from 'd3-format';
 
-const { ASPECT_RATIO } = require('./layout');
+const LayoutInjector = require('inject-loader!./layout');
+
 
 describe('points module', () => {
-    let layoutMock = proxyquire('./layout', {
-        './domain': {
-            tickSelector: () => {
-                return {
-                    tickValues: [5, 10, 15],
-                    tickFormat: format('d')
-                };
+    let Layout;
+
+    beforeEach(() => {
+        Layout = LayoutInjector({
+            './domain': {
+                tickSelector: function () {
+                    return {
+                        tickValues: [5, 10, 15],
+                        tickFormat: format('d')
+                    };
+                }
             }
-        }
+        });
     });
 
     it('Should return the width and height with the predefined ASPECT_RATIO', () => {
-        let layout = layoutMock.layoutSelector({
+        let layout = Layout.layoutSelector({
             ui: {
                 width: 200,
                 windowWidth: 600
@@ -24,8 +28,7 @@ describe('points module', () => {
         });
 
         expect(layout.width).toEqual(200);
-        expect(layout.height).toEqual(200 * ASPECT_RATIO);
+        expect(layout.height).toEqual(200 * Layout.ASPECT_RATIO);
         expect(layout.windowWidth).toEqual(600);
     });
 });
-