+++ assets/src/scripts/floodData.spec.js
@@ -1,4 +1,4 @@
-const proxyquire = require('proxyquireify')(require);
+const FloodDataInjector = require('inject-loader!./floodData');
 
 describe('flood_data module', () => {
 
@@ -19,7 +19,7 @@ describe('flood_data module', () => {
                     }
                 };
                 spyOn(ajaxMock, 'get').and.callThrough();
-                floodData = proxyquire('./floodData', {'./ajax': ajaxMock});
+                floodData = FloodDataInjector({'./ajax': ajaxMock});
             });
 
             it('url contains the siteno', () => {
@@ -47,7 +47,7 @@ describe('flood_data module', () => {
                         return getPromise;
                     }
                 };
-                floodData = proxyquire('./floodData', {'./ajax': ajaxMock});
+                floodData = FloodDataInjector({'./ajax': ajaxMock});
             });
 
             it('On failed response return an empty feature list', () => {
@@ -75,7 +75,7 @@ describe('flood_data module', () => {
                     }
                 };
                 spyOn(ajaxMock, 'get').and.callThrough();
-                floodData = proxyquire('./floodData', {'./ajax': ajaxMock});
+                floodData = FloodDataInjector({'./ajax': ajaxMock});
             });
 
             it('url contains the siteno', () => {
@@ -104,7 +104,7 @@ describe('flood_data module', () => {
                         return getPromise;
                     }
                 };
-                floodData = proxyquire('./floodData', {'./ajax': ajaxMock});
+                floodData = FloodDataInjector({'./ajax': ajaxMock});
             });
 
             it('On failed response return an empty feature list', () => {