+++ assets/src/scripts/statisticsData.spec.js
@@ -1,4 +1,4 @@
-let proxyquire = require('proxyquireify')(require);
+const StatisicsDataInjector = require('inject-loader!./statisticsData');
 
 
 describe('statisticsData', () => {
@@ -21,7 +21,7 @@ describe('statisticsData', () => {
                 }
             };
             spyOn(ajaxMock, 'get').and.callThrough();
-            statisticsData = proxyquire('./statisticsData', {'./ajax': ajaxMock});
+            statisticsData = StatisicsDataInjector({'./ajax': ajaxMock});
         });
 
         it('Gets a full year of statistical data', () => {
@@ -51,7 +51,7 @@ describe('statisticsData', () => {
                 }
             };
             spyOn(ajaxMock, 'get').and.callThrough();
-            statisticsData = proxyquire('./statisticsData', {'./ajax': ajaxMock});
+            statisticsData = StatisicsDataInjector({'./ajax': ajaxMock});
         });
 
         it('Gets a full year of statistical data', () => {