+++ assets/src/scripts/store/index.spec.js
@@ -1,6 +1,8 @@
-let proxyquire = require('proxyquireify')(require);
+import { Actions } from './index';
+const { sortedParameters } = require('../models');
 
-const { Actions } = require('./index');
+const StoreInjector = require('inject-loader!./index');
+
 
 describe('Redux store', () => {
 
@@ -78,7 +80,7 @@ describe('Redux store', () => {
                 spyOn(modelsMock, 'queryWeatherService').and.callThrough();
                 mockDispatch = jasmine.createSpy('mockDispatch');
                 mockGetState = jasmine.createSpy('mockGetState').and.returnValue(TEST_STATE);
-                store = proxyquire('./index', {'../models': modelsMock});
+                store = StoreInjector({'../models': modelsMock});
                 store.configureStore();
             });
 
@@ -116,7 +118,7 @@ describe('Redux store', () => {
                 spyOn(modelsMock, 'queryWeatherService').and.callThrough();
                 mockDispatch = jasmine.createSpy('mockDispatch');
                 mockGetState = jasmine.createSpy('mockGetState').and.returnValue(TEST_STATE);
-                store = proxyquire('./index', {'../models': modelsMock});
+                store = StoreInjector({'../models': modelsMock});
                 store.configureStore();
             });
 
@@ -143,13 +145,14 @@ describe('Redux store', () => {
                 modelsMock = {
                     getTimeSeries: function () {
                         return getTimeSeriesPromise;
-                    }
+                    },
+                    sortedParameters
                 };
 
                 spyOn(modelsMock, 'getTimeSeries').and.callThrough();
                 mockDispatch = jasmine.createSpy('mockDispatch');
                 mockGetState = jasmine.createSpy('mockGetState').and.returnValue(TEST_STATE);
-                store = proxyquire('./index', {'../models': modelsMock});
+                store = StoreInjector({'../models': modelsMock});
                 store.configureStore();
 
                 spyOn(store.Actions, 'addTimeSeriesLoading');
@@ -217,13 +220,14 @@ describe('Redux store', () => {
                 modelsMock = {
                     getTimeSeries: function () {
                         return getTimeSeriesPromise;
-                    }
+                    },
+                    sortedParameters
                 };
 
                 spyOn(modelsMock, 'getTimeSeries').and.callThrough();
                 mockDispatch = jasmine.createSpy('mockDispatch');
                 mockGetState = jasmine.createSpy('mockGetState').and.returnValue(TEST_STATE);
-                store = proxyquire('./index', {'../models': modelsMock});
+                store = StoreInjector({'../models': modelsMock});
                 store.configureStore();
             });
 
@@ -256,7 +260,7 @@ describe('Redux store', () => {
                 spyOn(modelsMock, 'getTimeSeries').and.callThrough();
                 mockDispatch = jasmine.createSpy('mockDispatch');
                 mockGetState = jasmine.createSpy('mockGetState').and.returnValue(TEST_STATE);
-                store = proxyquire('./index', {'../models': modelsMock});
+                store = StoreInjector({'../models': modelsMock});
 
                 spyOn(store.Actions, 'addTimeSeriesLoading');
                 spyOn(store.Actions, 'removeTimeSeriesLoading');
@@ -303,7 +307,7 @@ describe('Redux store', () => {
                 spyOn(modelsMock, 'getPreviousYearTimeSeries').and.callThrough();
                 mockDispatch = jasmine.createSpy('mockDispatch');
                 mockGetState = jasmine.createSpy('mockGetState').and.returnValue(TEST_STATE);
-                store = proxyquire('./index', {'../models': modelsMock});
+                store = StoreInjector({'../models': modelsMock});
                 spyOn(store.Actions, 'addTimeSeriesLoading');
                 spyOn(store.Actions, 'removeTimeSeriesLoading');
             });
@@ -354,7 +358,7 @@ describe('Redux store', () => {
                 spyOn(modelsMock, 'getPreviousYearTimeSeries').and.callThrough();
                 mockDispatch = jasmine.createSpy('mockDispatch');
                 mockGetState = jasmine.createSpy('mockGetState').and.returnValue(TEST_STATE);
-                store = proxyquire('./index', {'../models': modelsMock});
+                store = StoreInjector({'../models': modelsMock});
 
                 spyOn(store.Actions, 'addTimeSeriesLoading');
                 spyOn(store.Actions, 'removeTimeSeriesLoading');
@@ -397,7 +401,7 @@ describe('Redux store', () => {
                 };
                 spyOn(statisticsDataMock, 'fetchSiteStatistics').and.callThrough();
                 mockDispatch = jasmine.createSpy('mockDispatch');
-                store = proxyquire('./index',{'../statisticsData': statisticsDataMock});
+                store = StoreInjector({'../statisticsData': statisticsDataMock});
                 store.configureStore();
                 spyOn(store.Actions, 'addMedianStats');
                 spyOn(store.Actions, 'toggleTimeSeries');
@@ -436,7 +440,7 @@ describe('Redux store', () => {
                 mockDispatch = jasmine.createSpy('mockDispatch');
                 mockGetState = jasmine.createSpy('mockGetState');
 
-                store = proxyquire('./index', {'../models': modelsMock});
+                store = StoreInjector({'../models': modelsMock});
                 spyOn(store.Actions, 'addTimeSeriesLoading');
                 spyOn(store.Actions, 'removeTimeSeriesLoading');
             });
@@ -565,7 +569,7 @@ describe('Redux store', () => {
                     })
                 }));
 
-                store = proxyquire('./index', {'../models': modelsMock});
+                store = StoreInjector({'../models': modelsMock});
                 spyOn(store.Actions, 'addTimeSeriesLoading');
                 spyOn(store.Actions, 'removeTimeSeriesLoading');
             });
@@ -636,7 +640,7 @@ describe('Redux store', () => {
                 spyOn(floodDataMock, 'fetchFloodFeatures').and.callThrough();
                 spyOn(floodDataMock, 'fetchFloodExtent').and.callThrough();
                 mockDispatch = jasmine.createSpy('mockDispatch');
-                store = proxyquire('./index', {'../floodData': floodDataMock});
+                store = StoreInjector({'../floodData': floodDataMock});
 
             });
 
@@ -692,7 +696,7 @@ describe('Redux store', () => {
                 spyOn(floodDataMock, 'fetchFloodFeatures').and.callThrough();
                 spyOn(floodDataMock, 'fetchFloodExtent').and.callThrough();
                 mockDispatch = jasmine.createSpy('mockDispatch');
-                store = proxyquire('./index', {'../floodData': floodDataMock});
+                store = StoreInjector({'../floodData': floodDataMock});
 
             });