+++ assets/src/scripts/tooltips.spec.js
@@ -1,6 +1,5 @@
-const { select } = require('d3-selection');
-
-const { appendTooltip } = require('./tooltips');
+import { select } from 'd3-selection';
+import { appendTooltip } from './tooltips';
 
 
 describe('tooltips', () => {