+++ CHANGELOG.md
@@ -26,6 +26,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 not have to be coerced into a time series.
 - Embedded hydrograph is time zone aware.
 - Scrolling added to the 'provisional statement' in IFrame embed
+- The Browserify build tooling was replaced with Webpack.
 
 ### Fixed
 - A bug with the graph watermark intercepting mouseover events driving the tooltips was fixed.